requirements.txt
# Python Backend Dependencies - SURF Customer Feedback Agent
# ============================================================

# CrewAI - Multi-Agent Framework (Updated for Python 3.13 compatibility)
crewai>=0.201.0
crewai-tools>=0.8.3

# LangChain Core Dependencies
langchain>=0.1.20
langchain-openai>=0.1.8
langchain-community>=0.0.38
openai>=1.35.0

# Database - PostgreSQL (Updated for easier installation)
psycopg[binary]>=3.1.8
sqlalchemy>=2.0.23
alembic>=1.13.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.3.3

# HTTP and API clients
httpx==0.27.0
requests==2.32.3

# Integrations
slack-sdk==3.27.2
notion-client==2.2.1

# FastAPI (for optional API endpoints)
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Utilities
python-dateutil==2.9.0
colorama==0.4.6

# Development and testing
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.1.0
ipython==8.25.0